requirements.txt
# Dependencias esenciales para Ballers App
streamlit>=1.44.0
sqlalchemy>=2.0.0
pandas>=2.2.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.1

# Google APIs
google-api-python-client>=2.150.0
google-auth>=2.30.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0

# UI Components
streamlit-option-menu>=0.3.0

# Text processing
unidecode>=1.3.0

# Plotting (si usas plotly en el código)
plotly>=5.0.0